Renamed class. Creating a statement works
